added system.commandline and small refactor to app instantiation
